refactor: dedicated package for equalizer
