+++ app/src/main/java/com/cappielloantonio/tempo/ui/fragment/SettingsFragment.java
@@ -2,13 +2,6 @@ package com.cappielloantonio.tempo.ui.fragment;
 
 import static com.google.android.material.internal.ViewUtils.hideKeyboard;
 
-import android.app.Activity;
-import android.content.Context;
-import android.content.ComponentName;
-import android.content.Intent;
-import android.content.ServiceConnection;
-import android.media.audiofx.AudioEffect;
-import android.net.Uri;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -16,35 +9,9 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.OptIn;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.app.AppCompatDelegate;
-import androidx.appcompat.widget.Toolbar;
-import androidx.core.os.LocaleListCompat;
-import androidx.core.view.ViewCompat;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentTransaction;
-import androidx.lifecycle.ViewModelProvider;
-import androidx.media3.common.util.UnstableApi;
-import androidx.navigation.NavController;
-import androidx.navigation.NavOptions;
-import androidx.navigation.fragment.NavHostFragment;
-import androidx.preference.EditTextPreference;
-import androidx.preference.ListPreference;
-import androidx.preference.Preference;
-import androidx.preference.PreferenceCategory;
-import androidx.preference.PreferenceFragmentCompat;
-import androidx.preference.SwitchPreference;
 import androidx.fragment.app.Fragment;
 
 import com.cappielloantonio.tempo.R;
-import com.cappielloantonio.tempo.databinding.FragmentAlbumCatalogueBinding;
-import com.cappielloantonio.tempo.databinding.FragmentSettingsBinding;
-import com.cappielloantonio.tempo.helper.ThemeHelper;
-import com.cappielloantonio.tempo.interfaces.DialogClickCallback;
-import com.cappielloantonio.tempo.interfaces.ScanCallback;
-import com.cappielloantonio.tempo.service.EqualizerManager;
-import com.cappielloantonio.tempo.service.MediaService;
 import com.cappielloantonio.tempo.databinding.FragmentSettingsBinding;
 import com.cappielloantonio.tempo.ui.activity.MainActivity;
 import com.cappielloantonio.tempo.util.Preferences;
